[#5] refac: frontend 구조 개선
